Create README.md
This algorithm (for a Master-postgraduate exam)performs a series of analysis and visualisation operations on a directed graph (a network), created using arc and node data read from a data set (CSV file). 
The result is to calculate and visualise various properties of the graph, analyse communities within the graph and visualise the result with detailed graphs.